+++ processor/src/test/resources/expected/com/example/inject/BasicInjectModelDaggerModule.java
@@ -1,0 +1,12 @@
+package com.example.inject;
+
+import dagger.Binds;
+import dagger.Module;
+import javax.annotation.Generated;
+
+@Generated("arez.processor.ArezProcessor")
+@Module
+public interface BasicInjectModelDaggerModule {
+  @Binds
+  BasicInjectModel bindComponent(Arez_BasicInjectModel component);
+}